refactor(renderer): extract the speaker mix into SpeakerRenderStage::mix_channels
Step 5, the core move: the whole per-channel speaker loop (direct beds
one-hot, per-band VBAP objects, crossover split, per-sample gain slew,
monitoring outputs) relocates verbatim from `render_frame` into
`SpeakerRenderStage::mix_channels`, fed by a `SpeakerStageFrame` of
frame-scoped borrows and returning `SpeakerStageDiagnostics`.
`fill_band_gains` moves with it. `render_frame` keeps, in order: the
`first_render` swap, the mix call, the topology-guard drop, the frame
counter, `finalize_output`, and the clip/auto-gain section (control
writes stay caller-side by design). `channel_states` (ramps + slew)
remains shared and is passed in — see the module doc for the one-mix-
pass-per-frame invariant the cascaded mode must respect.

Bit-exact: full workspace suite incl. golden and null tests unchanged.
